% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_AssayObject}
\alias{as_AssayObject}
\title{as_AssayObject}
\usage{
as_AssayObject(object)
}
\arguments{
\item{object}{Currently supported is RCDT and Giotto (spatEnrObj) output}
}
\description{
Convert deconvolution output object to a TransferPrediction object
}
\examples{
# as_AssayObject(object)
}
