% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{as_GiottoObject}
\alias{as_GiottoObject}
\title{as_GiottoObject}
\usage{
as_GiottoObject(object, assay = "Spatial", slot = "counts", ...)
}
\arguments{
\item{object}{Seurat Object}

\item{assay}{Seurat assay in \code{object}}

\item{slot}{Seurat slot in \code{object} and \code{assay}}

\item{...}{Arguments passed to \code{Giotto::createGiottoObject}}
}
\description{
Convert Seurat Visium object to a GiottoObject
}
\examples{
# as_GiottoObject(object)
}
