% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_Reference}
\alias{as_Reference}
\title{as_Reference}
\usage{
as_Reference(object, refdata, assay = "RNA", slot = "counts", ...)
}
\arguments{
\item{object}{Seurat object}

\item{refdata}{Column with cell type annotation in \code{object}}

\item{assay}{Seurat assay in \code{object}}

\item{slot}{Seurat slot in \code{object} and \code{assay}}

\item{...}{Arguments passed to \code{spacexr::Reference}}
}
\description{
Convert Seurat single cell object for spacexr
}
\examples{
# as_Reference(object)
}
