% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_SingleCellExperiment}
\alias{as_SingleCellExperiment}
\title{as_SingleCellExperiment}
\usage{
as_SingleCellExperiment(object, ...)
}
\arguments{
\item{object}{Seurat Visium object}

\item{...}{Arguments passed to \code{spatialPreprocess}}
}
\description{
Convert Seurat Visium object for BayesSpace. Currently only works for single
sample objects.
}
\examples{
# as_SingleCellExperiment(object)
}
