% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{as_spatEnrObj}
\alias{as_spatEnrObj}
\title{as_spatEnrObj}
\usage{
as_spatEnrObj(
  object,
  slot = "data",
  ignore = c("max", "unassigned"),
  method = "DWLS",
  name = method,
  spat_unit = "cell",
  feat_type = "rna",
  ...
)
}
\arguments{
\item{object}{Seurat \code{AssayObject}}

\item{slot}{Seurat slot in \code{object}}

\item{ignore}{Features to be ignored}

\item{method}{Name of the method used to generate \code{object}}

\item{name}{Name of this enrichment object.}

\item{spat_unit}{Giotto spatial unit.}

\item{feat_type}{Giotto feature type.}

\item{...}{Additional arguments passed to \code{Giotto::new("spatEnrObj")}}
}
\description{
Convert \code{Seurat::AssayObject} to a \code{Giotto::spatEnrObj}
}
\examples{
# giotto_enrichment <- as_spatEnrObj(x$predictions)
# giotto_object <- as_GiottoObject(x)
# giotto_object <- set_spatial_enrichment(gobject = giotto_object,
#                                         spatenrichment = giotto_enrichment)
# spatDeconvPlot(giotto_object, radius = 150)
# 
}
