% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_SpatialRNA}
\alias{as_SpatialRNA}
\title{as_SpatialRNA}
\usage{
as_SpatialRNA(object, assay = "Spatial", slot = "counts", ...)
}
\arguments{
\item{object}{Seurat Visium object}

\item{assay}{Seurat assay in \code{object}}

\item{slot}{Seurat slot in \code{object} and \code{assay}}

\item{...}{Arguments passed to \code{spacexr::SpatialRNA}}
}
\description{
Convert Seurat Visium object for spacexr
}
\examples{
# as_SpatialRNA(object)
}
