% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{calculate_giotto_spatial_correlation_cell_type}
\alias{calculate_giotto_spatial_correlation_cell_type}
\title{calculate_giotto_spatial_correlation_cell_type}
\usage{
calculate_giotto_spatial_correlation_cell_type(
  gobject,
  cell_types,
  features = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{cell_types}{Cell type of interest}

\item{features}{Features of interest. If \code{NULL}, use all.}
}
\description{
Calculates spatial correlation of feature and cell type
}
\examples{
#
}
