% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_nmf_gse}
\alias{calculate_nmf_gse}
\title{calculate_nmf_gse}
\usage{
calculate_nmf_gse(
  obj,
  sig_set,
  method = c("fisher", "perm"),
  perm_n = 1000,
  scale = TRUE,
  alternative = c("two.sided", "less", "greater"),
  min_nmf_features = 100,
  summary_function = mean,
  rank,
  k,
  prefix,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_nmf}}.}

\item{sig_set}{List of signatures, read by \code{\link{read_signatures}}.}

\item{method}{Method to calculate p-values}

\item{perm_n}{If \code{method="perm"}, number of permutations}

\item{scale}{If \code{method="perm"}, scale NMF basis}

\item{alternative}{If \code{method="perm"}, direction to test}

\item{min_nmf_features}{Minimum number of features for each NMF cluster.
Default means that top 50 genes associated with cluster are tested for 
enrichment.}

\item{summary_function}{If \code{method="perm"}, function to summarize 
gene signature scores}

\item{rank}{Number of clusters (the one used in \code{\link{cluster_nmf}})}

\item{k}{Features of rank to be written (must be a single k, not a range)}

\item{prefix}{Prefix of output files}

\item{verbose}{Verbose output}
}
\description{
Correlate gene signatures with NMF clusters
}
\examples{
calculate_nmf_gse
}
