% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{calculate_spatial_correlation}
\alias{calculate_spatial_correlation}
\title{Spatial Correlation}
\usage{
calculate_spatial_correlation(
  obj,
  image = NULL,
  slot = "data",
  feature.1,
  feature.2 = rownames(obj)
)
}
\arguments{
\item{obj}{Seurat Object}

\item{image}{Passed to the GetTissueCoordinates function}

\item{slot}{Passed to the FetchData function}

\item{feature.1}{First feature}

\item{feature.2}{Second feature(s)}
}
\value{
spatial correlation
}
\description{
This function provides the spatial correlation of two features. If both are the
same, this corresponds to Moran's I.
}
