% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{classify_cell_type_regions}
\alias{classify_cell_type_regions}
\title{classify_cell_type_regions}
\usage{
classify_cell_type_regions(object, cell_type = "Tumor", cutoff = NULL)
}
\arguments{
\item{object}{Seurat object containing integrated}

\item{cell_type}{Cell-type of interest}

\item{cutoff}{Deconvolution cutoff for core membership}
}
\description{
Used to classify cell-types as core, periphery, diffused, other
after integration
}
\examples{
classify_cell_type_regions()
}
