% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_check_file_list}
\alias{cli_check_file_list}
\title{cli_check_file_list}
\usage{
cli_check_file_list(file, check_exists = TRUE)
}
\arguments{
\item{file}{Input .list file, alternatively a string
with filenames separated by :}

\item{check_exists}{Check input files}
}
\description{
Internal function for the command line interface
}
\examples{
cli_check_file_list()
}
