% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_check_lib_ids}
\alias{cli_check_lib_ids}
\title{cli_check_lib_ids}
\usage{
cli_check_lib_ids(reference_list)
}
\arguments{
\item{reference_list}{List of \code{Seurat} objects}
}
\description{
Internal function for the command line interface
}
\examples{
cli_check_lib_ids
}
