% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_bayesspace}
\alias{cluster_bayesspace}
\title{cluster_bayesspace}
\usage{
cluster_bayesspace(x, test_num_clusters = seq(2, 16), num_clusters = NULL, ...)
}
\arguments{
\item{x}{Object, converted by \code{\link{as_SingleCellExperiment}}}

\item{test_num_clusters}{Range of number of clusters to be tested (\code{qs}).}

\item{num_clusters}{Optional picked number}

\item{...}{Additional paramters passed to \code{BayesSpace::spatialCluster}}
}
\description{
Cluster BayesSpace data
}
\examples{
#cluster_bayesspace()
}
