% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{cluster_integrated}
\alias{cluster_integrated}
\title{cluster_integrated}
\usage{
cluster_integrated(integrated, regressout, scale = NULL, force,
  plot_umap = TRUE, serialize = TRUE, prefix, verbose = FALSE)
}
\arguments{
\item{integrated}{Seurat object containing integrated 
SpatialTranscriptomics and scRNA-seq data}

\item{regressout}{Regressout these features.}

\item{scale}{Scale data. If \code{NULL} use recommendation.}

\item{force}{Recalculate, even when serialized objects are available}

\item{plot_umap}{Generate UMAP plot of integrated data}

\item{serialize}{Serialize output objects}

\item{prefix}{Prefix of output files}

\item{verbose}{Verbose Seurat output}
}
\description{
Cluster integrated data (obtained by \code{\link{integrate_spatial}}.
}
\examples{
cluster_integrated()
}
