% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_prediction_strength}
\alias{cluster_prediction_strength}
\title{cluster_prediction_strength}
\usage{
cluster_prediction_strength(obj1, obj2, col1 = NULL, col2 = NULL,
  hejpeg1 = NULL, hejpeg2 = NULL, assay = "Spatial", prefix)
}
\arguments{
\item{obj1}{Object, clustered by \code{\link{cluster_spatial}}.}

\item{obj2}{Object, clustered by \code{\link{cluster_spatial}}.}

\item{col1}{\code{meta.data} column containing cluster labels of 
\code{obj1}. Default use first available.}

\item{col2}{\code{meta.data} column containing cluster labels of 
\code{obj2}. Default use first available.}

\item{hejpeg1}{Jpeg file of H&E for \code{obj1}}

\item{hejpeg2}{Jpeg file of H&E for \code{obj2}}

\item{assay}{Name of the assay corresponding to the initial input data.}

\item{prefix}{Prefix of output files}
}
\description{
Calculates the cluster consistency across two technical 
replicates. Shows the fraction of spot pairs clustered together
for each spot in both samples.
}
\examples{
cluster_prediction_strength
}
