% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{cluster_reference}
\alias{cluster_reference}
\title{cluster_reference}
\usage{
cluster_reference(integrated, resolution = 0.8, sub = FALSE,
  idents = NULL, features = NULL, force, plot_umap = TRUE,
  serialize = TRUE, prefix, verbose = FALSE, ...)
}
\arguments{
\item{integrated}{Seurat object containing integrated 
SpatialTranscriptomics and scRNA-seq data}

\item{resolution}{Cluster resolution}

\item{sub}{Flag indicating a subset of the data was provided}

\item{idents}{Use idents from a serialized Seurat object or as provided}

\item{features}{Use only the specified features for clustering}

\item{force}{Recalculate, even when serialized objects are available}

\item{plot_umap}{Output UMAP plot}

\item{serialize}{Serialize output objects}

\item{prefix}{Prefix of output files}

\item{verbose}{Verbose Seurat output}

\item{...}{Additional parameters passed to \code{Seurat::FindClusters}}
}
\description{
Cluster the scRNA-seq reference in the integrated data 
(obtained by \code{\link{integrate_spatial}}.
}
\examples{
cluster_reference()
}
