% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_sc3}
\alias{cluster_sc3}
\title{cluster_sc3}
\usage{
cluster_sc3(
  obj,
  rank,
  variable_features = FALSE,
  max_features = NULL,
  force,
  serialize = TRUE,
  prefix,
  ...
)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_spatial}}.}

\item{rank}{Number of clusters}

\item{variable_features}{If \code{TRUE}, only use variable features}

\item{max_features}{Reduce runtime by only using the top 
\code{max_features} features}

\item{...}{Additional parameters passed to the \code{sc3} function.}
}
\description{
Performes SC3 clustering
}
\examples{
cluster_sc3
}
