% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_spatial}
\alias{cluster_spatial}
\title{cluster_spatial}
\usage{
cluster_spatial(ndata, resolution = 0.8, dims = 1:30, verbose = TRUE)
}
\arguments{
\item{ndata}{Object, read by \code{\link{read_spatial}}.}

\item{resolution}{Seurat cluster resolution}

\item{dims}{PCs to use}

\item{verbose}{Verbose Seurat output}
}
\description{
Normalize spatial transcriptomics Seurat object
}
\examples{
cluster_spatial()
}
