% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{combine_assayobjects}
\alias{combine_assayobjects}
\title{combine_assayobjects}
\usage{
combine_assayobjects(
  x,
  min_fraction = 0.05,
  drop_zero = TRUE,
  ignore_unassigned = TRUE,
  labels = names(x)
)
}
\arguments{
\item{x}{list of \code{AssayObject}s}

\item{ignore_unassigned}{Ignore the unassigned category (not all methods
support them - currently only rctd_multi)}

\item{labels}{Labels of \code{x}, usually the method identifier}
}
\description{
Combines multiple cell type prediction into one AssayObject, keeps all predictions
}
\examples{
combine_assayobjects()
}
