% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{deconvolute_simulation}
\alias{deconvolute_simulation}
\title{deconvolute_simulation}
\usage{
deconvolute_simulation(simulation, references, min_features, regressout,
  resolution, idents, max_markers = NULL, plot_qc = TRUE,
  min_fraction = 0.01, verbose = FALSE)
}
\arguments{
\item{simulation}{True cluster proportions as obtained by \code{\link{simulate_spatial}}}

\item{references}{List of Seurat scRNA-seq reference datasets}

\item{min_features}{Remove cells or spots with small number of detected genes}

\item{regressout}{Regressout these features.}

\item{resolution}{Cluster resolution}

\item{idents}{Use idents from a serialized Seurat object or as provided}

\item{max_markers}{Use only top markers}

\item{plot_qc}{Plot concordance heatmap and scatter plot}

\item{min_fraction}{Minimum deconvoluted fraction to call cluster detected}

\item{verbose}{Verbose Seurat output}
}
\description{
Deconvolute simulated SpatialTranscriptomics data
}
\examples{
deconvolute_simulation()
}
