% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute_spatial}
\alias{deconvolute_spatial}
\title{deconvolute_spatial}
\usage{
deconvolute_spatial(obj_spatial, sig, write = TRUE, prefix, ...)
}
\arguments{
\item{obj_spatial}{Seurat object containing SpatialTranscriptomics data}

\item{sig}{Signature for deconvolution}

\item{write}{Write output data to text and PDF files}

\item{prefix}{Prefix of output files}

\item{...}{Additional parameters passed to \code{\link{plot_spots}}}
}
\description{
Simulate SpatialTranscriptomics data
}
\examples{
deconvolute_spatial()
}
