% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{enhance_deconvolve}
\alias{enhance_deconvolve}
\title{enhance_deconvolve}
\usage{
enhance_deconvolve(
  object,
  num_subspots = 6,
  max_cell_types = 7,
  num_iter = 10,
  stroke = NA,
  crop = FALSE,
  pt.size.factor = 0.4,
  image.alpha = 1,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{num_subspots}{Split spot into n subspots}

\item{max_cell_types}{Too avoid plotting too many cell types, we can limit them here.
Set to -1 to turn off.}

\item{num_iter}{Number optimization iterations}

\item{stroke}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{crop}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{pt.size.factor}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{image.alpha}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{...}{Additional parameters passed to \code{\link{Seurat::SingleSpatialPlot}}.}
}
\description{
Enhances deconvoluted Spatial data
}
\examples{
#enhance_deconvolve()
}
