% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{evaluate_simulation}
\alias{evaluate_simulation}
\title{evaluate_simulation}
\usage{
evaluate_simulation(simulation, deconvolution, plot_qc = TRUE,
  min_fraction = 0.01)
}
\arguments{
\item{simulation}{True cluster proportions as obtained by \code{\link{simulate_spatial}}}

\item{deconvolution}{Deconvoluted cluster proportions from \code{\link{deconvolute_spatial}}}

\item{plot_qc}{Generate concordance heatmap and scatter plots}

\item{min_fraction}{Minimum deconvoluted fraction to call cluster detected}
}
\description{
Evalue deconvolution of simulated SpatialTranscriptomics data
}
\examples{
evaluate_simulation()
}
