% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{export_nmf_loupe}
\alias{export_nmf_loupe}
\title{export_nmf_loupe}
\usage{
export_nmf_loupe(obj, rank, k, libs, labels = NULL, prefix)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_nmf}}.}

\item{rank}{Number of clusters (the one used in \code{\link{cluster_nmf}})}

\item{k}{Features of rank to be written (must be a single k, not a range)}

\item{libs}{Library ids, must be stored in \code{obj$library}}

\item{labels}{Optional \code{character(n)} with sample labels}

\item{prefix}{Prefix of output files}
}
\description{
Output NMF clustering to a CSV file loadable in Loupe
}
\examples{
export_nmf_loupe
}
