% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{export_snn_loupe}
\alias{export_snn_loupe}
\title{export_snn_loupe}
\usage{
export_snn_loupe(obj, prefix)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_nmf}}.}

\item{prefix}{Prefix of output files}
}
\description{
Output SNN clustering to a CSV file loadable in Loupe
}
\examples{
export_snn_loupe
}
