% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{find_cluster_neighborhoods}
\alias{find_cluster_neighborhoods}
\title{Cluster neighborhoods}
\usage{
find_cluster_neighborhoods(obj, max.dist = 3, fun.aggregate = mean)
}
\arguments{
\item{obj}{Seurat Object}

\item{max.dist}{Defines neighborhood as maximum distance in number
of spots}

\item{fun.aggregate}{Aggregate cluster value of spot values}
}
\value{
output pairwise cluster neighborhood overlaps
}
\description{
This function provides the neighborhood of clusters
}
