% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{find_markers}
\alias{find_markers}
\title{find_markers}
\usage{
find_markers(
  obj,
  references,
  resolution,
  max_markers = NULL,
  merged = FALSE,
  write = TRUE,
  force = FALSE,
  serialize = TRUE,
  prefix,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat object containing reference data obtained by
\code{\link{cluster_reference}}}

\item{references}{List of Seurat scRNA-seq reference datasets}

\item{resolution}{Cluster resolution}

\item{max_markers}{Use only top markers}

\item{merged}{Flag indicating whether clusters were already merged}

\item{write}{Write output text files and PDFs}

\item{force}{Recalculate, even when serialized objects are available}

\item{serialize}{Serialize output objects}

\item{prefix}{Prefix of output files}

\item{verbose}{Verbose Seurat output}

\item{...}{Additional parameters passed to \code{Seurat::FindMarkers}}
}
\description{
Integrate SpatialTranscriptomics data with (matched) scRNA-seq data
}
\examples{
find_markers()
}
