% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{find_nearest_neighbors}
\alias{find_nearest_neighbors}
\title{find_nearest_neighbors}
\usage{
find_nearest_neighbors(object, split.by = "library")
}
\arguments{
\item{object}{Seurat object containing integrated 
spatial data}

\item{split.by}{Split \code{object} by this feature.}
}
\description{
Shows the correlation of spots across multiple samples,
after integration
}
\examples{
find_nearest_neighbors()
}
