% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{find_nn_spatial}
\alias{find_nn_spatial}
\title{find_nn_spatial}
\usage{
find_nn_spatial(obj_spatial, obj_ref, k = 20, write = TRUE, prefix,
  ...)
}
\arguments{
\item{obj_spatial}{Seurat object containing SpatialTranscriptomics data}

\item{obj_ref}{Seurat object containing reference data}

\item{k}{k-nearest neighbors}

\item{write}{Write output data to text and PDF files}

\item{prefix}{Prefix of output files}

\item{...}{Additional parameters passed to \code{\link{plot_features}}}
}
\description{
Simulate SpatialTranscriptomics data
}
\examples{
find_nn_spatial()
}
