% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{find_signature_means}
\alias{find_signature_means}
\title{find_signature_means}
\usage{
find_signature_means(obj, features, slot = "data", label = NULL, fun = mean)
}
\arguments{
\item{obj}{Seurat object or a normalized count matrix (row genes, col samples)}

\item{features}{\code{character(n)} of features}

\item{slot}{If \code{obj} is a Seurat object, the slot name}

\item{label}{If not \code{NULL}, a \code{obj@meta.data} column name
to store.}

\item{fun}{Function to summary signature}

\item{...}{Arguments passed to \code{\link{plot_spots}}.}
}
\description{
A simple alternative to Seurat's \code{AddModuleScore}, simply
taking the signature mean without additional normalization
}
\examples{
find_signature_means()
}
