% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{import_loupe}
\alias{import_loupe}
\title{import_loupe}
\usage{
import_loupe(obj, file)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_nmf}}.}

\item{file}{CSV file loadable in Loupe}
}
\value{
Seurat object with annotation added.
}
\description{
Impurt annotation from a CSV file loadable in Loupe
}
\examples{
import_loupe
}
