% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_spatial}
\alias{normalize_spatial}
\title{normalize_spatial}
\usage{
normalize_spatial(
  obj,
  nfeatures = 2500,
  scale = TRUE,
  center = TRUE,
  correct_umi = TRUE,
  method = c("sctransform", "sctransform2", "seurat2", "scran"),
  backend_method = "poisson",
  feature_filter = TRUE,
  cell_cycle_score = TRUE,
  regressout = NULL,
  assay = "Spatial",
  serialize = TRUE,
  prefix,
  ...
)
}
\arguments{
\item{obj}{Object, read by \code{\link{read_spatial}}.}

\item{nfeatures}{Number of variable features}

\item{scale}{Use \code{Seurat::ScaleData} to scale data.}

\item{center}{Use \code{Seurat::ScaleData} to center data.}

\item{correct_umi}{Currently only supported by \code{sctransform}}

\item{method}{Use \code{sctransform} to normalize and scale data,
or use \code{scran} or Seurat 2 style normalization and scaling.}

\item{backend_method}{Method used by the backend. Currently only used by 
\code{sctransform} and \code{sctransform2} since the default assay is kept unchanged.}

\item{feature_filter}{If feature meta data contains 'included' column, only use}

\item{cell_cycle_score}{Use \code{CellCycleScoring} to add S/G1/G2M scores}

\item{regressout}{Regressout these features.}

\item{assay}{Name of the assay corresponding to the initial input data.}

\item{serialize}{Automatically serialize object}

\item{prefix}{Prefix of output files}

\item{...}{Additional parameters passed to the normalization 
function.}
}
\description{
Normalize spatial transcriptomics Seurat object
}
\examples{
normalize_spatial()
}
