% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{plot_cellphone}
\alias{plot_cellphone}
\title{Plot cell-cell interaction analysis using CellPhoneDB}
\usage{
plot_cellphone(obj, id = 1, cellphone_outpath)
}
\arguments{
\item{obj}{Seurat Object}

\item{id}{Line of significant pair in significant_means_ranked_spatial.csv file}

\item{cellphone_outpath}{Path with CellPhoneDB output}
}
\value{
ggplot object
}
\description{
This function allows you to read reranked CellPhoneDB output and plot it.
}
\examples{
plot_cellphone
}
