% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{plot_correlation_heatmap}
\usage{
plot_correlation_heatmap(objs, features, sample_labels = NULL,
  feature_labels = NULL, average_nn = FALSE)
}
\arguments{
\item{objs}{List of Seurat objects}

\item{features}{Features to plot}

\item{sample_labels}{\code{character(n)} with sample labels}

\item{feature_labels}{\code{character(n)} with feature labels. 
If \code{NULL}, just use indices.}

\item{average_nn}{Include the 4 nearest neighbors (top, bottom,
right, left) in the correlation}
}
\description{
Given a list of Seurat objects, plots the fraction of available genes
for each signature in a GMT file
}
\examples{
plot_correlation_heatmap
}
