% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nmf}
\alias{plot_nmf}
\title{plot_nmf}
\usage{
plot_nmf(obj, libs, hejpegs, labels = NULL, rank, prefix,
  subdir = "nmf", width = 10, png = FALSE, plot_he = TRUE,
  plot_ranks = TRUE, plot_qc = TRUE, assay = "Spatial", ...)
}
\arguments{
\item{obj}{Seurat object that contains a \code{NMF} object after
running \code{\link{cluster_nmf}}}

\item{libs}{Library ids, must be stored in \code{obj$library}}

\item{hejpegs}{Matching HE jpegs to \code{libs}}

\item{labels}{Optional \code{character(n)} with sample labels}

\item{rank}{Number of clusters}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{width}{Output PDF width}

\item{png}{Create, in addition to PDF, PNG files}

\item{plot_he}{Plot for each library the HE jpeg}

\item{plot_ranks}{Plot diagnostics about all tested ranks}

\item{plot_qc}{Plot QC statistics about all tested ranks}

\item{assay}{Name of the assay corresponding to the initial input data.}

\item{...}{Additional parameters passed to \code{\link{plot_features}}}
}
\description{
Basic plots of a saved \code{NMF} object
}
\examples{
plot_nmf()
}
