% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nmf_gse}
\alias{plot_nmf_gse}
\title{plot_nmf_gse}
\usage{
plot_nmf_gse(obj, sig, binwidth = 0.025, rank, prefix, subdir = "nmf",
  method = "", width = 10, png = FALSE)
}
\arguments{
\item{obj}{Results of \code{\link{calculate_nmf_gse}}}

\item{sig}{The name of the signature to plot. Must match a row name
in \code{obj}.}

\item{binwidth}{P-value histogram binwidth}

\item{rank}{Number of clusters}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{method}{Method that was used. Only used in output filenames.}

\item{width}{Output PDF width}

\item{png}{Create, in addition to PDF, PNG files}
}
\description{
Plots the association of a gene signature with a NMF clustering
}
\examples{
plot_nmf_gse
}
