% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prediction_summary}
\alias{plot_prediction_summary}
\title{plot_prediction_summary}
\usage{
plot_prediction_summary(
  object,
  feature = "spot_class",
  label = NULL,
  label_integration_method = "default",
  prefix,
  subdir = "advanced",
  pdf = FALSE,
  png = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Name of the feature that contains the prediction summary}

\item{label}{Label of single cell dataset}

\item{label_integration_method}{Label of integration method}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{pdf}{Create PDF image files}

\item{png}{Create PNG image files}

\item{...}{Arguments passed to \code{Seurat::SpatialDimPlot}.}
}
\description{
Plots a categorical prediction summary in a Seurat object on HE slides.
}
\examples{
plot_prediction_summary()
}
