% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictions}
\alias{plot_predictions}
\title{plot_predictions}
\usage{
plot_predictions(
  object,
  predictions = NULL,
  label = NULL,
  label_integration_method = "default",
  prefix,
  subdir = "he",
  png = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{predictions}{object with transfer predictions}

\item{label}{Label of single cell dataset}

\item{label_integration_method}{Label of integration method}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{png}{Create, in addition to PDF, PNG files}

\item{...}{Arguments passed to \code{plot_features}.}

\item{width}{Output PDF width}

\item{height}{Output PDF height}

\item{ncol}{Number of columns to plot genes for multi-page plots}

\item{nrow}{Number of rows to plot genes for multi-page plots}
}
\description{
Plots predictions in a Seurat object on HE slides. Supports multi-page
plot when lots of features are requested.
}
\examples{
plot_predictions()
}
