% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_qc_read}
\alias{plot_qc_read}
\title{plot_qc_read}
\usage{
plot_qc_read(object, prefix, assay)
}
\arguments{
\item{object}{Seurat object}

\item{prefix}{Prefix of output files}

\item{assay}{Name of the assay corresponding to the initial input data.}
}
\description{
Plot QC after reading a Spatial read counts
}
\examples{
plot_qc_read
}
