% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sc3}
\alias{plot_sc3}
\title{plot_sc3}
\usage{
plot_sc3(
  object,
  libs,
  labels = NULL,
  rank,
  prefix,
  subdir = "sc3",
  width = 10,
  pdf = FALSE,
  png = FALSE,
  plot_he = TRUE,
  assay = "Spatial",
  ...
)
}
\arguments{
\item{object}{Seurat object that contains \code{SC3} information after
running \code{\link{cluster_sc3}}}

\item{libs}{Library ids, must be stored in \code{object$library}}

\item{labels}{Optional \code{character(n)} with sample labels}

\item{rank}{Number of clusters}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{width}{Output PDF width}

\item{pdf}{Create PDF image files}

\item{png}{Create PNG image files}

\item{plot_he}{Plot for each library the HE jpeg}

\item{assay}{Name of the assay corresponding to the initial input data.}

\item{...}{Additional parameters passed to \code{\link{plot_features}}}
}
\description{
Basic plots of saved \code{SC3} information
}
\examples{
plot_sc3()
}
