% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scoloc}
\alias{plot_scoloc}
\title{plot_scoloc}
\usage{
plot_scoloc(celltrek_object, directed = FALSE)
}
\arguments{
\item{celltrek_object}{\code{celltrek} object to plot co-localization for}

\item{directed}{Should the graph be directed or not (default: FALSE)}
}
\description{
Co-localization plot for cell types from \code{celltrek} integration
}
\examples{
plot_scoloc(celltrek)
}
