% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_signatures_fake_bulk}
\alias{plot_signatures_fake_bulk}
\title{plot_signatures_fake_bulk}
\usage{
plot_signatures_fake_bulk(
  objs,
  gmt,
  assay = "Spatial",
  log_trans = TRUE,
  labels = NULL,
  plot_pairs = TRUE,
  plot_bar = TRUE,
  plot_heatmaps = TRUE
)
}
\arguments{
\item{objs}{Seurat objects with SpatialTranscriptomics data}

\item{gmt}{GMT file with gene signatures or gene signature read by
\code{\link{read_signatures}}}

\item{assay}{Assay to extract counts from}

\item{log_trans}{Transform counts into log-space.}

\item{labels}{Optional \code{character(n)} vector with labels}

\item{plot_pairs}{Plot pairwise correlations of all samples}

\item{plot_bar}{Plot median counts per M for each sample}

\item{plot_heatmap}{Plot heatmaps of all signatures in \code{gmt}}
}
\description{
Plots signatures defined in a GMT file and applied to list of
Seurat objects, ignores spatial information. This is useful for
a 30,000ft look at the data, similar to bulk
}
\examples{
plot_signatures_fake_bulk()
}
