% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_signatures_nmf}
\alias{plot_signatures_nmf}
\title{plot_signatures_nmf}
\usage{
plot_signatures_nmf(
  object,
  gmt,
  gmt_name = NULL,
  rank,
  prefix,
  subdir = "nmf",
  width = 10
)
}
\arguments{
\item{object}{Seurat object that contains a \code{NMF} object after
running \code{\link{cluster_nmf}}}

\item{gmt}{GMT file with gene signatures or gene signature read by
\code{\link{read_signatures}}}

\item{rank}{Number of clusters}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{width}{Output PDF width}

\item{png}{Create, in addition to PDF, PNG files}
}
\description{
Plot correlation of gene signatures with basis of an \code{NMF} object
}
\examples{
plot_signatures_nmf()
}
