% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spots}
\alias{plot_spots}
\title{plot_spots}
\usage{
plot_spots(
  x,
  labels = scales::percent,
  labels_title = "Contribution",
  undetected_NA = TRUE,
  undetected_cutoff = 0.001,
  na.rm = TRUE,
  max_quantile = 0.975,
  reorder_clusters = TRUE,
  palette = "viridis",
  palette_inverse = FALSE,
  alpha = 0.85,
  size = 0.9,
  trans = NULL,
  limits = NULL
)
}
\arguments{
\item{x}{\code{data.frame} with spot coordinates, value and cluster name.}

\item{labels}{transformation of the labels, e.g. \code{scales::percent}}

\item{labels_title}{Title, shown in the legend}

\item{undetected_NA}{Remove spots with low value by setting them to \code{NA}}

\item{undetected_cutoff}{Define low value by this cutoff}

\item{na.rm}{Hide spots with NA values}

\item{max_quantile}{Cap values at this quantile to avoid that outliers.
Ignored if \code{limits} is not \code{NULL}.     
define the color scheme}

\item{reorder_clusters}{Reorder clusters alphabetically}

\item{palette}{Color palette. Can be two colors in format "low:high",
brewer_single_hue_x (x = red, green, blue, orange, gray, purple) for
single hue palettes from colorbrewer2.org}

\item{palette_inverse}{Flip the low and high colors in \code{palette}}

\item{alpha}{Transparency of spots}

\item{size}{Size of spot dots}

\item{trans}{Transform values, currently only log2 is implemented when this 
parameter is set to a different value than \code{NULL}.}

\item{limits}{Optional limits for continuous scores}
}
\description{
Plots spots in a Seurat object on HE slides
}
\examples{
plot_spots()
}
