% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_violin}
\alias{plot_violin}
\title{plot_violin}
\usage{
plot_violin(
  obj,
  features,
  cells = NULL,
  zero_offset = NULL,
  pt_size = 0.25,
  slot = "data",
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{\code{character(n)} of features to be plotted}

\item{cells}{Plot only specified cells}

\item{zero_offset}{If zero values were set to a low value, specify the offset here.}

\item{pt_size}{Size of overlayed dots}

\item{slot}{Slot to pull feature data for}

\item{...}{Arguments passed to \code{Seurat::VlnPlot}.}
}
\description{
Violin plot of a Seurat object
}
\examples{
plot_violin()
}
