% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{read_signatures}
\alias{read_signatures}
\title{read_signatures}
\usage{
read_signatures(gmt, obj = NULL, max_nchar_title = 25)
}
\arguments{
\item{gmt}{GMT file with gene signatures. Can also be a CSV file with
columns \code{gene} and \code{cluster} as generated by \code{\link{find_markers}}}

\item{obj}{Seurat object, used to filter signatures for non-available genes}

\item{max_nchar_title}{Truncate signature name}
}
\description{
Read signatures defined in a GMT file
}
\examples{
read_signatures()
}
