% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{read_spaceranger_deconvolution}
\alias{read_spaceranger_deconvolution}
\title{read_spaceranger_deconvolution}
\usage{
read_spaceranger_deconvolution(
  obj,
  filtered_feature_bc_matrix_dir,
  deconv_dir = file.path(filtered_feature_bc_matrix_dir, "deconvolution")
)
}
\arguments{
\item{obj}{Seurat object}

\item{filtered_feature_bc_matrix_dir}{Path to SpaceRanger filtered matrix}

\item{deconv_dir}{Path to SpaceRanger \code{spatial} directory}
}
\description{
Add 10X SpaceRanger deconvolution into Seurat object
}
\examples{
read_visium()
}
