% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_spaceranger_metrics}
\alias{read_spaceranger_metrics}
\title{read_spaceranger_metrics}
\usage{
read_spaceranger_metrics(
  filtered_feature_bc_matrix_dir,
  metrics_file = file.path(filtered_feature_bc_matrix_dir, "metrics_summary.csv")
)
}
\arguments{
\item{filtered_feature_bc_matrix_dir}{Path to SpaceRanger filtered matrix}

\item{metrics_file}{Path to SpaceRanger \code{metrics_summary.csv} file}
}
\description{
Parse 10X SpaceRanger metrics
}
\examples{
#read_spaceranger_metrics
}
