% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_spaceranger_probe_set}
\alias{read_spaceranger_probe_set}
\title{read_spaceranger_probe_set}
\usage{
read_spaceranger_probe_set(file)
}
\arguments{
\item{file}{Path to SpaceRanger probe sets file}
}
\description{
Parse 10X SpaceRanger probe sets
}
\examples{
#read_spaceranger_probe_set
}
