% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_spatial}
\alias{read_spatial}
\title{read_spatial}
\usage{
read_spatial(
  file,
  sampleid,
  mt_pattern = regex_mito(),
  rp_pattern = regex_ribo(),
  min_features = 300,
  min_spots = 2,
  required_features = NULL,
  transpose = FALSE,
  barcodes = NULL,
  image = NULL,
  slice = sampleid,
  downsample_prob = NULL,
  hybrid_reference_prefixes = c("hg19", "mm10"),
  gtf = NULL,
  assay = "Spatial",
  serialize = TRUE,
  prefix
)
}
\arguments{
\item{file}{File or data containing Spatial read counts}

\item{sampleid}{\code{character(1)} with sampleid}

\item{mt_pattern}{Regex to match mitochondria genes}

\item{rp_pattern}{Regex to match ribosomal genes}

\item{min_features}{Minimum number of features to include spot}

\item{min_spots}{Minimum number of spots to include gene}

\item{required_features}{Add those features even if not present or
passing \code{min_spots}}

\item{transpose}{Transpose input matrix}

\item{barcodes}{Optionally, barcodes for all spots}

\item{image}{Optional \code{VisiumV1} object containing image information.
Can also be a JPEG for Spatial Transcriptomics data.}

\item{slice}{Optional name for the stored \code{image}}

\item{downsample_prob}{Downsample input matrix. Requires \code{DropletUtils}.}

\item{hybrid_reference_prefixes}{For dataset aligned to hybrid references, the
the feature name prefixes. First one reserved for human (for non-human,
hybrid references, this just means that some downstream tools might not work).}

\item{gtf}{Optional GTF for including feature meta data like alternative gene ids.}

\item{assay}{Name of the assay corresponding to the initial input data.}

\item{serialize}{Automatically serialize object}

\item{prefix}{Prefix of output files}
}
\description{
Read spatial transcriptomics into Seurat object
}
\examples{
read_spatial()
}
