% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_visium}
\alias{read_visium}
\title{read_visium}
\usage{
read_visium(filtered_feature_bc_matrix_dir, spatial_dir,
  tissue_positions_file = "tissue_positions_list.csv", ...)
}
\arguments{
\item{filtered_feature_bc_matrix_dir}{Path to SpaceRanger filtered matrix}

\item{spatial_dir}{Path to SpaceRanger \code{spatial} directory}

\item{tissue_positions_file}{File providing the coordindates}

\item{...}{Arguments passed to \code{\link{read_spatial}}}
}
\description{
Read 10X SpaceRanger data into Seurat object
}
\examples{
read_visium()
}
