% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{regex_ribo}
\alias{regex_ribo}
\title{regex_ribo}
\usage{
regex_ribo(hg = "hg19.", mm = "mm10.")
}
\arguments{
\item{hg}{Gene symbol prefix for human genes in hybrid references}

\item{mm}{Gene symbol prefix for mouse genes in hybrid references}
}
\description{
Regular expression to robustly find ribosomal genes
}
\examples{
regex_ribo()
}
