% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{set_idents_nmf}
\alias{set_idents_nmf}
\title{set_idents_nmf}
\usage{
set_idents_nmf(object, k, rank = NULL, stop_if_unavail = FALSE)
}
\arguments{
\item{object}{Object, clustered by \code{\link{cluster_nmf}}.}

\item{k}{Features of rank to be written (must be a single k, not a range)}

\item{rank}{Number of clusters (the one used in \code{\link{cluster_nmf}})}

\item{stop_if_unavail}{Dies when NMF clustering is not found.
If \code{NULL}, use first available.}
}
\value{
object with new \code{Idents}
}
\description{
Sets \code{Idents} to a discrete NMF clustering
}
\examples{
set_idents_nmf
}
