% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{set_idents_sc3}
\alias{set_idents_sc3}
\title{set_idents_sc3}
\usage{
set_idents_sc3(obj, k, rank = NULL, stop_if_unavail = FALSE)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_sc3}}.}

\item{k}{Features of rank to be written (must be a single k, not a range)}

\item{rank}{Number of clusters (the one used in \code{\link{cluster_sc3}})}

\item{stop_if_unavail}{Dies when SC3 clustering is not found.
If \code{NULL}, use first available.}
}
\value{
obj with new \code{Idents}
}
\description{
Sets \code{Idents} to SC3 clustering
}
\examples{
set_idents_sc3
}
