% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_spatial}
\alias{simulate_spatial}
\title{simulate_spatial}
\usage{
simulate_spatial(
  obj_single_cell,
  obj_spatial,
  clusters_per_spot,
  cells_per_spot,
  write = TRUE,
  prefix
)
}
\arguments{
\item{obj_single_cell}{Seurat object containing scRNA-seq reference}

\item{obj_spatial}{Seurat object containing SpatialTranscriptomics 
template data}

\item{clusters_per_spot}{Assume that many clusters contribute to each spot.}

\item{cells_per_spot}{Assume that many cells contribute to each spot.}

\item{write}{Write simulated data to text files}

\item{prefix}{Prefix of output files}
}
\description{
Simulate SpatialTranscriptomics data
}
\examples{
simulate_spatial()
}
