% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{write_nmf_features}
\alias{write_nmf_features}
\title{write_nmf_features}
\usage{
write_nmf_features(obj, rank, k, prefix)
}
\arguments{
\item{obj}{Object, clustered by \code{\link{cluster_nmf}}.}

\item{rank}{Number of clusters (the one used in \code{\link{cluster_nmf}})}

\item{k}{Features of rank to be written (must be a single k, not a range)}

\item{prefix}{Prefix of output files}
}
\description{
Output metagenes to a CSV file
}
\examples{
write_nmf_features
}
