% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{cellphone_for_seurat}
\alias{cellphone_for_seurat}
\title{cell-cell interaction analysis using method CellPhoneDB}
\usage{
cellphone_for_seurat(obj, orgdb, prefix, slot = "data", assay = NULL)
}
\arguments{
\item{obj}{Seurat Object}

\item{orgdb}{orgdb Object}

\item{prefix}{Prefix of output files}

\item{slot}{Specific information to pull
(i.e. counts, data, scale.data,...)}

\item{assay}{Name of assay to pull data from}
}
\value{
output tables for cell-cell interction inference
}
\description{
This function allows you to use an CellPhoneDB on Seurat Object
}
\examples{

}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{cell-cell}
\keyword{interactions}
\keyword{sequencing,}
\keyword{single}
