% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{enhance_deconvolve}
\alias{enhance_deconvolve}
\title{enhance_deconvolve}
\usage{
enhance_deconvolve(
  object,
  num_subspots = 6,
  max_cell_types = 7,
  cell_types = NULL,
  image = NULL,
  num_iter = 10,
  stroke = NA,
  crop = FALSE,
  pt.size.factor = 0.4,
  pt.alpha = NULL,
  image.alpha = 1,
  verbose = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{num_subspots}{Split spot into n subspots}

\item{max_cell_types}{Too avoid plotting too many cell types, we can limit them here.
Set to -1 to turn off. Ordered by prevalence if \code{cell_types} is \code{NULL}.}

\item{cell_types}{Only plot the specified cell types.}

\item{image}{Image to plot}

\item{num_iter}{Number optimization iterations}

\item{stroke}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{crop}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{pt.size.factor}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{pt.alpha}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{image.alpha}{\code{\link{Seurat::SingleSpatialPlot}} argument}

\item{verbose}{Verbose output}
}
\description{
Enhances deconvoluted Spatial data
}
\examples{
#enhance_deconvolve()
}
