% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{find_assayobject_consensus}
\alias{find_assayobject_consensus}
\title{find_assayobject_consensus}
\usage{
find_assayobject_consensus(
  x,
  min_fraction = 0.05,
  drop_zero = TRUE,
  ignore_unassigned = TRUE,
  labels = names(x),
  plot_correlations = FALSE
)
}
\arguments{
\item{x}{list of \code{AssayObject}s}

\item{min_fraction}{Set all probabilities smaller this value to 0}

\item{drop_zero}{Drop cell types never assigned}

\item{ignore_unassigned}{Ignore the unassigned category (not all methods
support them - currently only rctd_multi)}

\item{labels}{Labels of \code{x}, usually the method identifier}

\item{plot_correlations}{Plot heatmap of correlations across \code{x}}
}
\description{
Averages multiple cell type prediction into one AssayObject
}
\examples{
find_assayobject_consensus()
}
