% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{import_cellphone}
\alias{import_cellphone}
\title{Import cell-cell interaction analysis using CellPhoneDB}
\usage{
import_cellphone(obj, cellphone_outpath, orgdb)
}
\arguments{
\item{obj}{Seurat Object}

\item{cellphone_outpath}{CellPhoneDB output path}

\item{orgdb}{orgdb Object}
}
\value{
interaction pairs in GMT format
}
\description{
This function allows you to read CellPhoneDB output and rerank it based on local interactions.
}
