% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_features}
\alias{plot_features}
\title{plot_features}
\usage{
plot_features(
  obj,
  features,
  cells = NULL,
  undetected_NA = FALSE,
  plot_correlations = TRUE,
  plot_violin = TRUE,
  plot_map = FALSE,
  zero_offset = NULL,
  slot = "data",
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{\code{character(n)} of features to be plotted}

\item{cells}{Plot only specified cells}

\item{undetected_NA}{Remove spots with low value by setting them to \code{NA}}

\item{plot_correlations}{Plot pairwise correlations of feature scores}

\item{plot_violin}{Plot distribution of feature scores across idents}

\item{plot_map}{Plots features in \code{facet_wrap} for colorblind creatures.
Requires that features are \code{factors}, like cluster labels.}

\item{zero_offset}{If zero values were set to a low value, specify the offset here.}

\item{slot}{Slot to pull feature data for}

\item{...}{Arguments passed to \code{\link{plot_spots}}.}
}
\description{
Plots features in a Seurat object on HE slides
}
\examples{
plot_features()
}
