% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spatially_variable}
\alias{plot_spatially_variable}
\title{plot_spatially_variable}
\usage{
plot_spatially_variable(
  ndata,
  labels = NULL,
  spatial_features,
  method = "markvariogram",
  number_features = 80,
  prefix,
  subdir = "spatial_variation",
  width = 10,
  ncol = 4,
  nrow = 4,
  ...
)
}
\arguments{
\item{labels}{Optional \code{character(n)} vector with labels}

\item{spatial_features}{Results of\code{Seurat::SpatiallyVariableFeatures}.}

\item{method}{Method used to find spatially variable features}

\item{number_features}{Plot that many features.}

\item{prefix}{Output file prefix}

\item{subdir}{Put files in a subdirectory}

\item{width}{Output PDF width}

\item{ncol}{Number of columns to plot genes}

\item{nrow}{Number of rows to plot genes}

\item{...}{Arguments passed to \code{\link{plot_features}}}

\item{obj}{Seurat object}
}
\description{
Plots results of spatial differential expression analysis
}
\examples{
plot_spatially_variable
}
