% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_visium}
\alias{read_visium}
\title{read_visium}
\usage{
read_visium(
  filtered_feature_bc_matrix_dir,
  spatial_dir = file.path(filtered_feature_bc_matrix_dir, "spatial"),
  assay = "Spatial",
  probe_set = NULL,
  ...
)
}
\arguments{
\item{filtered_feature_bc_matrix_dir}{Path to SpaceRanger filtered matrix}

\item{spatial_dir}{Path to SpaceRanger \code{spatial} directory}

\item{assay}{Name of the assay corresponding to the initial input data.}

\item{probe_set}{Path to SpaceRanger probe set file. Only useful
if SpaceRanger was run without probe set filtering.}

\item{...}{Arguments passed to \code{\link{read_spatial}}}
}
\description{
Read 10X SpaceRanger data into Seurat object
}
\examples{
read_visium()
}
